Script agora guarda histórico de preços. Possível fazer busca por nome do jogo

+++ Trabalho-Final/busca_precos.sh
@@ -10,10 +10,14 @@ while getopts ":ugq" opt; do
 	# Busca todos os ID's dos jogos no banco de dados
 	game_IDs=$( cat name_ID_list.json | grep 'appid' | grep -oP '[^\D]*' )
 	
-	# Cria um arquivo para guardar os valores encontrados
-	# A ideia é depois substituir valores, não sobrescrever o arquivo
-	# mas por enquanto está assim
-	echo "" > name_ID_price_list.txt
+	# Cria um arquivo para guardar os valores encontrados, se o mesmo já não existir
+	if [ ! -f "name_ID_price_list.txt" ]; then
+		echo "" > name_ID_price_list.txt
+	fi
+
+	count=0
+
+	echo "Buscando jogos... ($count/$( wc -w <<< $game_IDs ))"
 		
 	# Para cada um dos ID's dos jogos no banco de dados
 	for ID in $game_IDs; do
@@ -21,10 +25,14 @@ while getopts ":ugq" opt; do
 		# Faz a busca pelo jogo na API
 		data=$( wget -qO - https://store.steampowered.com/api/appdetails?appids=$ID )
 
+		count=$( expr $count + 1 )
+
 		# Se não retornou nada, parte para o próximo ID. Muitos não são encontrados mesmo
 		# Não sabemos a qualidade da API com que estamos lidando
 		if [ -z "$data" ]; then
-			echo "Falha o buscar jogo. ID: $ID. Jogo não encontrado"
+			echo -en "\r"
+			echo "Falha ao buscar jogo. ID: $ID. Jogo não encontrado"
+			echo -n "Buscando jogos... ($count/$( wc -w <<< $game_IDs ))"
 			continue
 		fi
 		
@@ -36,7 +44,9 @@ while getopts ":ugq" opt; do
 
 		# Checa se houve sucesso na hora de buscar o jogo
 		if [ $status == "false" ]; then
-			echo "Falha ao buscar jogo. ID: $ID. Success status == False"
+			echo -en "\r"
+			echo -e "Falha ao buscar jogo. ID: $ID. Success status == False"
+			echo -n "Buscando jogos... ($count/$( wc -w <<< $game_IDs ))"
 			continue
 		else
 
@@ -49,31 +59,104 @@ while getopts ":ugq" opt; do
 
 			nome=$( echo $data | grep -o 'name":"[^"]*' --color=always | sed -n '1p' | cut -b 19- )
 			if [ -z "$nome" ]; then
-				echo "Falha o buscar jogo. ID: $ID. Nome não encontrado"
+				echo -en "\r"
+				echo "Falha ao buscar jogo. ID: $ID. Nome não encontrado"
+				echo -n "Buscando jogos... ($count/$( wc -w <<< $game_IDs ))"
 				continue
 			fi
 			nome=${nome:: -6}
 
 			preco=$( echo $data | grep -o 'final":[0-9]*' --color=always | sed -n '1p' | cut -b 19- )
 			if [ -z "$preco" ]; then
-				echo "Falha o buscar jogo. ID: $ID. Preço não encontrado"
+				echo -en "\r"
+				echo "Falha ao buscar jogo. ID: $ID. Preço não encontrado"
+				echo -n "Buscando jogos... ($count/$( wc -w <<< $game_IDs ))"
 				continue
 			fi
 			preco=${preco:: -6}
 
-			# Insere infos do jogo no arquivo
-			echo $nome  >> name_ID_price_list.txt
-			echo $ID    >> name_ID_price_list.txt
-			echo $preco >> name_ID_price_list.txt
-			echo ""		>> name_ID_price_list.txt
 
-			echo "Jogo atualizado com sucesso. ID: $ID, Nome: $nome"
+			tmp_id=$( grep -e "^$ID$" name_ID_price_list.txt )
+			tmp_preco=$( grep -A1 -e "^$ID$" name_ID_price_list.txt | tail -n 1 )
+
+			#echo $tmp_preco
+
+			# Se entrada ainda não existir, insere ela no arquivo. Caso contrário
+			# substitui o valor, se for menor
+			if [ -z "$tmp_id" ]; then
+				#Insere infos do jogo no arquivo
+				echo ""		>> name_ID_price_list.txt
+				echo $nome  >> name_ID_price_list.txt
+				echo $ID    >> name_ID_price_list.txt
+				echo $preco >> name_ID_price_list.txt
+			else
+
+				if [ "$tmp_preco" -gt "$preco" ]; then
+					# Sobrescreve as informações do jogo que acabou de ser buscado usando o sed
+					# pra substituir as linhas
+					# -0: Lê o arquivo inteiro para permitir substituição em múltiplas linhas
+					# -i: Permite edição para substituir
+					# -p: Pattern matching
+					# -e: Executa comando
+					perl -i -0pe "s/\n.*\n$ID\n.*\n/\n$nome\n$ID\n$preco\n/" name_ID_price_list.txt
+				fi
+			fi
+
+			echo -en "\r"
+			echo -e "Jogo atualizado com sucesso. ID: $ID, Nome: $nome"
+			echo -n "Buscando jogos... ($count/$( wc -w <<< $game_IDs ))"
 		fi
 	done
+	echo ""
 		#echo $game_IDs
       ;;
     q ) # Faz uma busca no banco de dados pelo menor preço do jogo. To be implemented
-	echo $2
+
+
+	if [ ! -f "name_ID_price_list.txt" ]; then
+		echo "Arquivo de histórico não existe"
+	fi
+
+	if [ -z "$2" ]; then
+		echo "Favor informar o nome do Jogo: $0 -q <Nome>"
+		exit
+	fi
+
+	# Busca quantidade de linhas que contém a string passada como input
+	num_matches=$( grep -c -ie "^.*$2.*$" name_ID_price_list.txt )
+
+	let num_matches=num_matches+1
+
+	# Se a consulta for muito pouco específica, vai retornar muitas entradas
+	# Seta valor máximo de entradas a serem mostradas como 10. Passível de modificação
+	if [ "$num_matches" -gt 10 ]; then
+		echo "Too many matches. Please be more specific with your search"
+		exit
+	fi
+
+	echo $num_matches "Matches Found:"
+
+	count=1
+	p="p"
+
+	# Para cada uma das entradas encontradas, buasca o nome, ID e preço do jogo
+	while [ ! "$count" -eq "$num_matches" ]; do
+
+		nome=$( grep -ie "^.*$2.*$" name_ID_price_list.txt | sed -n "$count p" )
+
+		id=$( grep -A1 -ie "^.*$nome.*$" name_ID_price_list.txt | tail -n 1 )
+
+		preco=$( grep -A1 -ie "^.*$id.*$" name_ID_price_list.txt | tail -n 1 )
+
+		echo ""
+		echo "Game: " $nome
+		echo "Steam ID: " $id
+		echo "Cheapest Recorded Price: R\$$(($preco/100)).$(($preco - 100*($preco/100)))"
+
+		let count=count+1
+
+	done
+
       ;;
     \? ) echo "Como usar: cmd [-u] [-g] [-q]"
 		 echo "Opção Inválida: $OPTARG" 1>&2
